Create ruleObjectsFailed from rules/compliant fields (#558)
* Create ruleObjectsFailed from rules/compliant fields

The response ruleObjectsFailed hasn't been reviewed on the backend for a
long time, and it returns the wrong rules. Currently it looks at
historical RuleResults (all of them!) for the system, and retrieves the
failing ones. This is NOT what we're interested on. We're
interested only on whether the rules for the *latest test results for
the current profiles* are compliant or not.

* Refactor names

* Tests for all helper methods

* Use lodash/flatten

* Use flatMap add add back remediationAvailable

* Only compute ruleObjectsFailed when remediations button is present