fix(CompliancePolicies): RHICOMPL-1128 Apollo error on refetch
Essential identifier was missing.

Here is the error:

Error writing result to store for query:
Store error: the application attempted to write an object with no provided id
but the store already contains an id of Profile:UUID for this object.
The selectionSet that was trying to be written is: ...