fix(EditPolicySystemsTab): RHICOMPL-1697 new rules alert
The new rules alert should only show when new tabs appear. This PR fixes
an issue where the alert appeared for new systems, despite not having
a valid policy type due to the OS version. This solution highlights the
performance issue with the rules tab requesting all benchmark rules.
This request is required to determine if the newly added system OS
versions have this policy type, so the alert is quite sluggish, on the
order of 1-2 seconds, to update when new systems are selected. I'm
hopeful we'll optimize the rules tab queries, which will also fix this
sluggish alert.

Signed-off-by: Andrew Kofink <akofink@redhat.com>