fix(Modal): RHICOMPL-2387 - Add modal with compliance class for scoped CSS (#1418)
Application specific CSS is now scoped on the platform, which means all our components will need to be within an element with the application ID as a class to have styles properly applied. The platform chrome does this already for the main part of our application, but not for modals. This adds a modal with the class predefined to use within compliance and have CSS styles available and similarly adds the compliance class to the Wizard (modal)